requirements.txt
playwright
ddddocr
python-telegram-bot
APScheduler
pytz
